requirements.txt
streamlit
groq
scikit-learn
faiss-cpu
python-dotenv
